beta 1.0

+++ src/app/components/stage/stage.component.ts
@@ -4,6 +4,21 @@ import { GameService } from 'src/app/services/game.service';
 import { Grid } from 'src/app/services/Grid';
 
 
+enum UserAction
+{
+  NO_ACTION,
+  ROTATE,
+  ROTATE_COUNTER,
+  PAUSE,
+  CONTINUE,
+  MOVE_RIGHT,
+  MOVE_LEFT,
+  STORE,
+  DROP,
+  SOFTDROP,
+  HARDDROP,
+}
+
 @Component({
   selector: 'app-stage',
   templateUrl: './stage.component.html',
@@ -11,23 +26,87 @@ import { Grid } from 'src/app/services/Grid';
 })
 export class StageComponent implements OnInit {
 
-  @HostListener("document:keypress", ["$event"])
-  onKeypress(e: KeyboardEvent){
-    switch(e.key){
+  currentAction: UserAction[] = [];
+
+  @HostListener("document:keydown", ["$event"])
+  @HostListener("document:keyup", ["$event"])
+  onKeyEvent(e: KeyboardEvent){
+
+    console.log(e);
+    let k = e.key;
+    if(k.length === 1){
+      k = k.toLowerCase();
+    }
+
+    const action = this.currentAction;
+    switch(k){
     case "w":
-      this.gameService.rotate();
+    case "x":
+    case "ArrowUp":
+      if(action.includes(UserAction.ROTATE)){
+        if(e.type === "keyup"){
+          action.splice(action.indexOf(UserAction.ROTATE), 1);
+        }
+      } else{
+        this.gameService.rotate();
+        action.push(UserAction.ROTATE);
+      }
+      break;
+    case "Control":
+    case "z":
+      if(action.includes(UserAction.ROTATE_COUNTER)){
+        if(e.type === "keyup"){
+          action.splice(action.indexOf(UserAction.ROTATE_COUNTER), 1);
+        }
+      } else{
+        this.gameService.rotate(false);
+        action.push(UserAction.ROTATE_COUNTER);
+      }
+      break;
+    case "Escape":
+    case "F1":
+      this.gameService.pause();
       break;
     case "d":
-      this.gameService.moveRight();
-      break;
-    case "a":
-      this.gameService.moveLeft();
+    case "ArrowRight":
+      
+      if(e.type === "keydown"){
+        this.gameService.moveRight();
+      }
       break;
     case "s":
-      this.gameService.drop();
+    case "ArrowDown":
+      if(action.includes(UserAction.SOFTDROP)){
+        if(e.type === "keyup"){
+          this.gameService.stopSoftDrop();
+          action.splice(action.indexOf(UserAction.SOFTDROP), 1);
+        }
+      } else{
+        this.gameService.startSoftDrop();
+        action.push(UserAction.SOFTDROP);
+      }
+      break;
+    case "a":
+    case "ArrowLeft":
+      if(e.type === "keydown"){
+        this.gameService.moveLeft();
+      }
+      break;
+    case "s":
+      break;
+    case "c":
+    case "Shift":
+      this.gameService.store();
       break;
     case " ":
-      this.gameService.store();
+      if(action.includes(UserAction.HARDDROP)){
+        if(e.type === "keyup"){
+          action.splice(action.indexOf(UserAction.HARDDROP), 1);
+        }
+      } else{
+        this.gameService.drop();
+        action.push(UserAction.HARDDROP);
+      }
       break;
     default:
     }
